Protect $ROLE_REQUIREMENTS with quotes.
This is important in case the filename contains spaces, but more
importantly, if no file is set, the test for file existence does not give
the desired outcome: [ -f ] evaluates to true/returns an exit code of 0,
hence [ -f $VAR ] with $VAR empty does the same. Protecting $VAR with
quotes has the desired behaviour.